template styling

+++ events/views.py
@@ -75,14 +75,15 @@ def attending(request):
 def delete(request):
     slug = request.GET.get('event')
     event= Events.objects.get(slug=slug)
-    if event.user  == request.user:
-        event.delete()
-        status='deleted'
-
-    if event.page.user == request.user:
-        event.delete()
-        status='deleted'
-
+    if event.user:
+        if event.user  == request.user:
+            event.delete()
+            status='deleted'
+    if event.page:
+        if event.page.user == request.user:
+            event.delete()
+            status='deleted'
+        
     return JsonResponse({'comment':status})
